DOC-1725 fix broken links, #6 of 8
In GSQL Docs version 3.6, an image was not found on the Defining A Graph Schema page. Interestingly, this image appears twice on the page, so the image is available. It was correctly referenced the first time; the second time there was a garbage filename.  I just fixed the file name.  This error was only in 3.6, not in newer versions.